+++ torch/cuda/memory.py
@@ -14,10 +14,10 @@ import torch
 from torch import _C
 
 from torch.types import Device
+from .._utils import _dummy_type
 from . import _get_device_index, _get_nvml_device_index, _lazy_init, is_initialized
 
 from ._memory_viz import memory as _memory, segments as _segments
-from .._utils import _dummy_type
 
 __all__ = [
     "caching_allocator_alloc",
